修正bug

+++ src/structs/EffectData.h
@@ -1,12 +1,12 @@
 #ifndef EFFECT_DATA_H
 #define EFFECT_DATA_H
 struct EffectData {
-    int speedEffect = 0;
-    int hitRateEffect = 0;
-    int pAttackEffect = 0;
-    int mAttackEffect = 0;
-    int pDefenseEffect = 0;
-    int mDefenseEffect = 0;
-    int focusEffect = 0;
+    double speedEffect = 0;
+    double hitRateEffect = 0;
+    double pAttackEffect = 0;
+    double mAttackEffect = 0;
+    double pDefenseEffect = 0;
+    double mDefenseEffect = 0;
+    double focusEffect = 0;
 };
 #endif
